Remove event listener if results are shown
Fixes a bug where it would still search on Enter even if results were shown, if the event listener was already activated.
eg. if you searched for "a" results would show, but "as" would redirect to search engine. However, if you deleted the "s" and pressed Enter it would still search for "as" even though results were shown for "a"